Move interface e fix namespaces
