+++ RuiSantos.ZocDoc.Core/InternalExtensions.cs
